KxSMBItemState support large size file

+++ KxSMBProvider.h
@@ -78,7 +78,7 @@ typedef void (^KxSMBBlockProgress)(KxSMBItem *item, long transferred);
 @interface KxSMBItemStat : NSObject
 @property(readonly, nonatomic, strong) NSDate *lastModified;
 @property(readonly, nonatomic, strong) NSDate *lastAccess;
-@property(readonly, nonatomic) long size;
+@property(readonly, nonatomic) long long size;
 @property(readonly, nonatomic) long mode;
 @end
